requirements.txt
pyserial==3.5
smllib==1.2
